removed all pertaining to synthetic data generatio
