+++ SEMIF.py
@@ -5,12 +5,11 @@ import hydra
 from hydra.utils import get_method, get_original_cwd, to_absolute_path
 from omegaconf import DictConfig, OmegaConf
 
-import develop_images
 import auto_sfm
+import develop_images
 import localize_plants  # Do not remove
 import remap_labels  # Do not remove
 import segment_vegetation  # Do not remove
-import synthesize
 
 log = logging.getLogger(__name__)
