+++ semif_utils/segment_utils.py
@@ -1,13 +1,9 @@
 import json
-from dataclasses import asdict, make_dataclass
-from datetime import datetime
 from pathlib import Path
-from pprint import pprint
 
 import cv2
 import numpy as np
 from dacite import Config, from_dict
-from omegaconf import DictConfig
 from scipy import ndimage as ndi
 from skimage import measure
 from skimage.color import label2rgb
@@ -16,15 +12,9 @@ from skimage.filters import rank
 from skimage.measure import label
 from skimage.morphology import disk
 from skimage.segmentation import watershed
-from tqdm import tqdm
 
-from semif_utils.datasets import (CUTOUT_PROPS, BatchMetadata, BBoxMetadata,
-                                  Cutout, CutoutProps, ImageData)
-from semif_utils.mongo_utils import Connect, _id_query, to_db
-from semif_utils.synth_utils import save_dataclass_json
-from semif_utils.utils import (apply_mask, clear_border, crop_cutouts,
-                               dilate_erode, get_site_id, get_upload_datetime,
-                               make_exg, make_exg_minus_exr, make_exr,
+from semif_utils.datasets import CUTOUT_PROPS, CutoutProps, ImageData
+from semif_utils.utils import (make_exg, make_exg_minus_exr, make_exr,
                                make_kmeans, make_ndi, otsu_thresh, parse_dict,
                                reduce_holes, rescale_bbox)
 
@@ -35,15 +25,15 @@ from semif_utils.utils import (apply_mask, clear_border, crop_cutouts,
 
 def get_siteinfo(imagedir):
     """Uses image directory to gather site specific information.
-            Agnostic to what relative path structure is used. As in it does
-            not matter whether parent directory of images is sitedir or "developed". 
+        Agnostic to what relative path structure is used. As in it does
+        not matter whether parent directory of images is sitedir or "developed".
 
-        Returns:
-            sitedir: developed image parent directory name
-            site_id: state id takend from sitedir
-        """
+    Returns:
+        sitedir: developed image parent directory name
+        site_id: state id takend from sitedir
+    """
     imagedir = Path(imagedir)
-    states = ['TX', 'NC', 'MD']
+    states = ["TX", "NC", "MD"]
     sitedir = [p for st in states for p in imagedir.parts if st in p][0]
     site_id = [st for st in states if st in sitedir][0]
     return sitedir, site_id
@@ -52,25 +42,26 @@ def get_siteinfo(imagedir):
 def get_image_meta(path):
     with open(path) as f:
         j = json.load(f)
-        imgdata = from_dict(data_class=ImageData,
-                            data=j,
-                            config=Config(check_types=False))
+        imgdata = from_dict(
+            data_class=ImageData, data=j, config=Config(check_types=False)
+        )
     return imgdata
 
 
 def save_cutout_json(cutout, cutoutpath):
-    cutout_json_path = Path(
-        Path(cutoutpath).parent,
-        Path(cutoutpath).stem + ".json")
-    with open(cutout_json_path, 'w') as j:
+    cutout_json_path = Path(Path(cutoutpath).parent, Path(cutoutpath).stem + ".json")
+    with open(cutout_json_path, "w") as j:
         json.dump(cutout, j, indent=4, default=str)
 
 
 def get_species_info(path, cls, default_species="grass"):
     with open(path) as f:
         spec_info = json.load(f)
-        spec_info = spec_info["species"][cls] if cls in spec_info[
-            "species"].keys() else default_species
+        spec_info = (
+            spec_info["species"][cls]
+            if cls in spec_info["species"].keys()
+            else default_species
+        )
     return spec_info
 
 
@@ -80,15 +71,12 @@ def get_species_info(path, cls, default_species="grass"):
 
 
 class GenCutoutProps:
-
     def __init__(self, mask):
-        """ Generate cutout properties and returns them as a dataclass.
-        """
+        """Generate cutout properties and returns them as a dataclass."""
         self.mask = mask
 
     def from_regprops_table(self, connectivity=2):
-        """Generates list of region properties for each cutout mask
-        """
+        """Generates list of region properties for each cutout mask"""
         labels = measure.label(self.mask, connectivity=connectivity)
         props = [measure.regionprops_table(labels, properties=CUTOUT_PROPS)]
         # Parse regionprops_table
@@ -102,7 +90,6 @@ class GenCutoutProps:
 
 
 class ClassifyMask:
-
     def otsu(self, vi):
         # Otsu's thresh
         vi_mask = otsu_thresh(vi)
@@ -117,7 +104,6 @@ class ClassifyMask:
 
 
 class VegetationIndex:
-
     def exg(self, img):
         exg_vi = make_exg(img, thresh=True)
         return exg_vi
@@ -138,9 +124,9 @@ class VegetationIndex:
 def get_image_meta(path):
     with open(path) as f:
         j = json.load(f)
-        imgdata = from_dict(data_class=ImageData,
-                            data=j,
-                            config=Config(check_types=False))
+        imgdata = from_dict(
+            data_class=ImageData, data=j, config=Config(check_types=False)
+        )
     return imgdata
 
 
@@ -154,19 +140,19 @@ def thresh_vi(vi, low=20, upper=100, sigma=2):
                                 "low" and "upper". Defaults to 2.
     """
     thresh_vi = np.where(vi <= 0, 0, vi)
-    thresh_vi = np.where((thresh_vi > low) & (thresh_vi < upper),
-                         thresh_vi * sigma, thresh_vi)
+    thresh_vi = np.where(
+        (thresh_vi > low) & (thresh_vi < upper), thresh_vi * sigma, thresh_vi
+    )
     return thresh_vi
 
 
 def seperate_components(mask):
-    """ Seperates multiple unconnected components in a mask
-        for seperate processing. 
+    """Seperates multiple unconnected components in a mask
+    for seperate processing.
     """
     # Store individual plant components in a list
     mask = mask.astype(np.uint8)
-    nb_components, output, _, _ = cv2.connectedComponentsWithStats(
-        mask, connectivity=8)
+    nb_components, output, _, _ = cv2.connectedComponentsWithStats(mask, connectivity=8)
     # Remove background component
     nb_components = nb_components - 1
     list_filtered_masks = []
@@ -199,7 +185,7 @@ def get_watershed(mask, disk1=1, grad1_thresh=12, disk2=10, lbl_fact=2.5):
 
 
 def species_info(speciesjson, default_species="grass"):
-    """Compares entries in user provided species map csv with those from a common 
+    """Compares entries in user provided species map csv with those from a common
        species data model (json). Uses 'get_close_matches' to get the best match.
        Is meant to create flexibility in how users are defining "species" in their
        maps.s
@@ -216,7 +202,7 @@ def species_info(speciesjson, default_species="grass"):
 
     # get species map dictionary unique to batch
 
-    spec_map = df.set_index('row').T.to_dict("records")[0]
+    spec_map = df.set_index("row").T.to_dict("records")[0]
     spec_map = eval(repr(spec_map).lower())
     spec_map_copy = spec_map.copy()
 
@@ -224,9 +210,7 @@ def species_info(speciesjson, default_species="grass"):
     species_data = read_json(speciesjson)
     common_names = []
     spec_idx = species_data["species"].keys()
-    common_name_list = [
-        species_data["species"][x]["common_name"] for x in spec_idx
-    ]
+    common_name_list = [species_data["species"][x]["common_name"] for x in spec_idx]
     # Get copy species map to update
     update_specmap = spec_map.copy()
