+++ semif_utils/utils.py
@@ -32,9 +32,10 @@ def read_json(path):
 
 def get_bbox_info(csv_path):
 
-    df = pd.read_csv(csv_path).drop(columns=['Unnamed: 0'])
-    bbox_dict = df.groupby(
-        by='imgname', sort=True).apply(lambda x: x.to_dict(orient='records'))
+    df = pd.read_csv(csv_path).drop(columns=["Unnamed: 0"])
+    bbox_dict = df.groupby(by="imgname", sort=True).apply(
+        lambda x: x.to_dict(orient="records")
+    )
     img_list = list(bbox_dict.keys())
     return bbox_dict, img_list
 
@@ -52,7 +53,7 @@ def creation_date(path_to_file):
     last modified if that isn't possible.
     See http://stackoverflow.com/a/39501288/1709587 for explanation.
     """
-    if platform.system() == 'Windows':
+    if platform.system() == "Windows":
         return os.path.getctime(path_to_file)
     else:
         stat = os.stat(path_to_file)
@@ -67,8 +68,7 @@ def creation_date(path_to_file):
 def get_upload_datetime(imagedir):
 
     creation_dt = creation_date(imagedir)
-    creation_dt = datetime.fromtimestamp(creation_dt).strftime(
-        '%Y-%m-%d_%H:%M:%S')
+    creation_dt = datetime.fromtimestamp(creation_dt).strftime("%Y-%m-%d_%H:%M:%S")
     return creation_dt
 
 
@@ -80,7 +80,7 @@ def parse_dict(props_tabl):
         new_val_entry = []
         if isinstance(val, np.ndarray) and val.shape[0] > 1:
             for i, v in enumerate(val):
-                new_val_entry.append({f'{key}_{i+1}': float(v)})
+                new_val_entry.append({f"{key}_{i+1}": float(v)})
             ndict[key] = new_val_entry
         else:
             ndict[key] = float(val)
@@ -143,10 +143,7 @@ def make_ndi(rgb_img):
     red = img[:, :, 0]
     gminr = green - red
     gplusr = green + red
-    gdivr = np.true_divide(gminr,
-                           gplusr,
-                           out=np.zeros_like(gminr),
-                           where=gplusr != 0)
+    gdivr = np.true_divide(gminr, gplusr, out=np.zeros_like(gminr), where=gplusr != 0)
     ndi = 128 * (gdivr + 1)
     # print("Max ndi: ", ndi.max())
     # print("Min ndi: ", ndi.min())
@@ -167,7 +164,7 @@ def rescale_bbox(box, imgshape):
     Args:
         box (dataclass): box metedata from bboxes from image metadata
         imgshape: np.ndarray: dimensions of the image to be scaled to (widt, height)
-    
+
     Returns:
         box (dataclass): box metadata with scaled/updated bbox
     """
@@ -206,11 +203,7 @@ def clean_mask(mask, kernel_size=3, iterations=1, dilation=True):
     return mask
 
 
-def dilate_erode(mask,
-                 kernel_size=3,
-                 dil_iters=5,
-                 eros_iters=3,
-                 hole_fill=True):
+def dilate_erode(mask, kernel_size=3, dil_iters=5, eros_iters=3, hole_fill=True):
     mask = mask.astype(np.float32)
 
     if int(kernel_size):
@@ -237,7 +230,8 @@ def reduce_holes(mask, min_object_size=1000, min_hole_size=1000):
     mask = mask.astype(np.bool8)
     # mask = measure.label(mask, connectivity=2)
     mask = morphology.remove_small_holes(
-        morphology.remove_small_objects(mask, min_hole_size), min_object_size)
+        morphology.remove_small_objects(mask, min_hole_size), min_object_size
+    )
     # mask = morphology.opening(mask, morphology.disk(3))
     return mask
 
@@ -267,8 +261,9 @@ def make_kmeans(exg_mask):
 
 def otsu_thresh(mask, kernel_size=(3, 3)):
     mask_blur = cv2.GaussianBlur(mask, kernel_size, 0).astype("uint16")
-    ret3, mask_th3 = cv2.threshold(mask_blur, 0, 255,
-                                   cv2.THRESH_BINARY + cv2.THRESH_OTSU)
+    ret3, mask_th3 = cv2.threshold(
+        mask_blur, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU
+    )
     return mask_th3
 
 
@@ -312,10 +307,131 @@ def crop_cutouts(img, add_padding=False):
     pil_crop_frground = foreground.crop(foreground.getbbox())
     array = np.array(pil_crop_frground)
     if add_padding:
-        pil_crop_frground = foreground.crop((
-            foreground.getbbox()[0] - 2,
-            foreground.getbbox()[1] - 2,
-            foreground.getbbox()[2] + 2,
-            foreground.getbbox()[3] + 2,
-        ))
+        pil_crop_frground = foreground.crop(
+            (
+                foreground.getbbox()[0] - 2,
+                foreground.getbbox()[1] - 2,
+                foreground.getbbox()[2] + 2,
+                foreground.getbbox()[3] + 2,
+            )
+        )
     return array
+
+
+# ------------------- Helper functions --------------------------------
+def bbox_iou(box1, box2):
+    box1 = torch.tensor([box1], dtype=torch.float)
+    box2 = torch.tensor([box2], dtype=torch.float)
+    iou = bops.box_iou(box1, box2)
+    return iou
+
+
+def get_img_bbox(x, y, imgshape):
+    pot_h, pot_w, _ = imgshape
+    x0, x1, y0, y1 = x, x + pot_w, y, y + pot_h
+    bbox = [x0, y0, x1, y1]  # top right corner, bottom left corner
+    return bbox
+
+
+def center2topleft(x, y, background_imgshape):
+    """Gets top left coordinates of an image from center point coordinate"""
+    back_h, back_w, _ = background_imgshape
+    y = y - int(back_h / 2)
+    x = x - int(back_w / 2)
+    return x, y
+
+
+def transform_position(points, imgshape, spread_factor):
+    """Applies random jitter factor to points and transforms them to top left image coordinates."""
+    y, x = points
+
+    x, y = x + random.randint(-spread_factor, spread_factor), y + random.randint(
+        -int(spread_factor / 3), int(spread_factor / 3)
+    )
+
+    x, y = center2topleft(x, y, imgshape)
+
+    return x, y
+
+
+def center_on_background(y, x, back_shape, fore_shape):
+    # pot positions and shape top left corner
+    back_h, back_w, _ = back_shape
+    fore_h, fore_w, _ = fore_shape
+    newx = int(((back_w - fore_w) / 2) + x)
+    newy = int(((back_h - fore_h) / 2) + y)
+    return newx, newy
+
+
+def img2RGBA(img):
+    alpha = np.sum(img, axis=-1) > 0
+    alpha = np.uint8(alpha * 255)
+    img = np.dstack((img, alpha))
+    return img
+
+
+class Point(object):
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+
+
+class Rect(object):
+    def __init__(self, p1, p2):
+        """Store the top, bottom, left and right values for points
+        p1 and p2 are the (corners) in either order
+        """
+        self.left = min(p1.x, p2.x)
+        self.right = max(p1.x, p2.x)
+        self.bottom = min(p1.y, p2.y)
+        self.top = max(p1.y, p2.y)
+
+
+def overlap(r1, r2):
+    """Overlapping rectangles overlap both horizontally & vertically"""
+    return range_overlap(r1.left, r1.right, r2.left, r2.right) and range_overlap(
+        r1.bottom, r1.top, r2.bottom, r2.top
+    )
+
+
+def range_overlap(a_min, a_max, b_min, b_max):
+    """Neither range is completely greater than the other"""
+    return (a_min <= b_max) and (b_min <= a_max)
+
+
+def dict_to_json(dic, path):
+    json_path = Path(path)
+    with open(json_path, "w") as j:
+        json.dump(dic, j, indent=4, default=str)
+
+
+def clean_data(data):
+    """Convert absolute pot and background paths to relative.
+    Takes the last two components of a path object for each.
+
+    Takes in and returns a dictionary of dataclass to be
+    stored in json and db.
+    """
+    data["background"]["background_path"] = "/".join(
+        Path(data["background"]["background_path"]).parts[-2:]
+    )
+    pots = data["pots"]
+    for pot in pots:
+        pot["pot_path"] = "/".join(Path(pot["pot_path"]).parts[-2:])
+
+    for cutout in data["cutouts"]:
+        cutout["cutout_path"] = "/".join(Path(cutout["cutout_path"]).parts[-2:])
+
+    return data
+
+
+def save_dataclass_json(data_dict, path):
+    json_path = Path(path)
+    with open(json_path, "w") as j:
+        json.dump(data_dict, j, indent=4, default=str)
+
+
+def get_cutout_dir(batch_dir, cutout_dir):
+    batch = Path(batch_dir).name
+    cutout_dir = Path(cutout_dir, batch)
+    return cutout_dir
